+++ api-server/api-server-common/src/storage/impls/postgres/transactional/read.rs
@@ -30,7 +30,7 @@ use crate::storage::{
         UtxoWithExtraInfo, block_aux_data::BlockAuxData,
     },
 };
-use crate::streaming::{StreamEvent, StreamEventId};
+use crate::streaming::{StreamEvent, StreamEventId, StreamEventReadError};
 use std::collections::BTreeMap;
 
 use common::chain::UtxoOutPoint;
@@ -453,7 +453,7 @@ impl ApiServerStorageRead for ApiServerPostgresTransactionalRo<'_> {
     async fn read_stream_events_after(
         &self,
         last_seen_id: StreamEventId,
-    ) -> Result<Vec<(StreamEventId, StreamEvent)>, ApiServerStorageError> {
+    ) -> Result<Vec<(StreamEventId, StreamEvent)>, StreamEventReadError> {
         let conn = QueryFromConnection::new(self.connection.as_ref().expect(CONN_ERR));
         let res = conn.read_stream_events_after(last_seen_id).await?;